Moved all AndroidBuilder implementations into android packages for docs
