+++ android/src/main/java/com/adsbynimbus/openrtb/android/AndroidRegsBuilder.java
@@ -1,6 +1,8 @@
-package com.adsbynimbus.openrtb.targeting;
+package com.adsbynimbus.openrtb.android;
 
 import androidx.annotation.NonNull;
+
+import com.adsbynimbus.openrtb.targeting.Regs;
 
 /**
  * Regs builder object with Androidx annotations for validation
