+++ android/src/main/java/com/adsbynimbus/openrtb/android/AndroidSourceBuilder.java
@@ -1,6 +1,8 @@
-package com.adsbynimbus.openrtb.targeting;
+package com.adsbynimbus.openrtb.android;
 
 import androidx.annotation.NonNull;
+
+import com.adsbynimbus.openrtb.targeting.Source;
 
 /**
  * Source builder implementation with Androidx annotations for validation
