+++ android/src/main/java/com/adsbynimbus/openrtb/android/ApiInt.java
@@ -1,6 +1,8 @@
-package com.adsbynimbus.openrtb.impression;
+package com.adsbynimbus.openrtb.android;
 
 import androidx.annotation.IntDef;
+
+import com.adsbynimbus.openrtb.impression.Api;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
