+++ android/src/main/java/com/adsbynimbus/openrtb/android/CreativeAttributeInt.java
@@ -1,6 +1,8 @@
-package com.adsbynimbus.openrtb.impression;
+package com.adsbynimbus.openrtb.android;
 
 import androidx.annotation.IntDef;
+
+import com.adsbynimbus.openrtb.impression.CreativeAttribute;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
